Add selectors/urls to Channels
