+++ webapp/channels/src/components/drafts/channel_draft/index.ts
@@ -6,7 +6,7 @@ import {connect} from 'react-redux';
 import {makeGetChannel} from 'mattermost-redux/selectors/entities/channels';
 import {getCurrentTeamId} from 'mattermost-redux/selectors/entities/teams';
 
-import {getChannelURL} from 'utils/utils';
+import {getChannelURL} from 'selectors/urls';
 
 import {GlobalState} from 'types/store';
